requirements.txt
mcp>=1.0.0
alpaca-py>=0.30.0
